chore: move tab description files to respective folders
