+++ resources/js/components/config/ConfigTabs.ts
