+++ resources/js/components/dashboard/DashboardTabs.ts
