+++ resources/js/components/settings/SettingsTabs.ts
